Split up `layers.marlin` into several files (#2292)
The marlin.py file was getting large, split it up.
